一旦、リファクタ
